LSTM_RCNN/requirements.txt
scipy==0.18.1

pandas==0.18.1

tensorflow==1.1.0

matplotlib==1.5.1

jupyter==1.0.0

scikit-learn==0.18

python-firebase
polyaxon

pymongo